refactor: alpinejs components
- this pr cleans up the app.js file and pulls all alpine logic into there own component js files making it much more organized and modular.
- this pr also converts the livewire post like and dislike buttons to alpinejs. this was done due to nesting components in livewire being bugged.
